Issue #1438, refactoring Pgp.armor (not ready)
