+++ extension/chrome/settings/modules/contacts.ts
@@ -22,6 +22,7 @@ import { Str, Url } from '../../../js/common/core/common.js';
 import { FetchKeyUI } from '../../../js/common/ui/fetch_key_ui.js';
 import { View } from '../../../js/common/view.js';
 import { Contact } from './../../../js/common/core/pgp';
+import { PgpArmor } from '../../../js/common/core/pgp/armor.js';
 
 View.run(class ContactsView extends View {
 
@@ -189,7 +190,7 @@ View.run(class ContactsView extends View {
       } else {
         pub = value;
       }
-      let { blocks } = Pgp.armor.detectBlocks(pub);
+      let { blocks } = PgpArmor.detectBlocks(pub);
       blocks = blocks.filter((b, i) => blocks.findIndex(f => f.content === b.content) === i); // remove duplicates
       if (!blocks.length) {
         await Ui.modal.warning('Could not find any new public keys.');
