+++ extension/chrome/settings/modules/my_key_update.ts
@@ -13,6 +13,7 @@ import { Attester } from '../../../js/common/api/attester.js';
 import { Api } from '../../../js/common/api/api.js';
 import { Url } from '../../../js/common/core/common.js';
 import { View } from '../../../js/common/view.js';
+import { PgpArmor } from '../../../js/common/core/pgp/armor.js';
 
 declare const openpgp: typeof OpenPGP;
 
@@ -21,7 +22,7 @@ View.run(class MyKeyUpdateView extends View {
   private readonly longid: string;
   private readonly showKeyUrl: string;
   private readonly inputPrivateKey = $('.input_private_key');
-  private readonly prvHeaders = Pgp.armor.headers('privateKey');
+  private readonly prvHeaders = PgpArmor.headers('privateKey');
   private primaryKi: KeyInfo | undefined;
 
   constructor() {
